requirements.txt
Django==3.2.7
django-compressor==2.4.1
django-crispy-forms==1.13.0
dj-database-url==0.5.0
requests==2.26.0
Pillow==8.3.2
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
django-debug-toolbar==3.2.2